+++ src/config/backend/EdFi.DmsConfigurationService.Backend.Mssql/Repositories/ApplicationRepository.cs
@@ -673,7 +673,7 @@ public class ApplicationRepository(
     }
 
     public async Task<ApplicationUpdateStateResult> GetApplicationUpdateState(
-        long applicationId,
+        int applicationId,
         string clientId
     )
     {
@@ -691,7 +691,7 @@ public class ApplicationRepository(
                 """;
             var application = await connection.QuerySingleOrDefaultAsync<(
                 string ApplicationName,
-                long VendorId,
+                int VendorId,
                 string ClaimSetName
             )?>(applicationSql, new { Id = applicationId, TenantId }, transaction);
 
@@ -706,7 +706,7 @@ public class ApplicationRepository(
                 WHERE ClientId = @ClientId AND ApplicationId = @ApplicationId;
                 """;
             var client = await connection.QuerySingleOrDefaultAsync<(
-                long Id,
+                int Id,
                 Guid ClientUuid,
                 bool IsApproved
             )?>(clientSql, new { ClientId = clientId, ApplicationId = applicationId }, transaction);
@@ -728,9 +728,9 @@ public class ApplicationRepository(
                 ),
             ];
 
-            long[] profileIds =
+            int[] profileIds =
             [
-                .. await connection.QueryAsync<long>(
+                .. await connection.QueryAsync<int>(
                     """
                     SELECT ProfileId FROM dmscs.ApplicationProfile
                     WHERE ApplicationId = @ApplicationId;
@@ -740,9 +740,9 @@ public class ApplicationRepository(
                 ),
             ];
 
-            long[] clientDataStoreIds =
+            int[] clientDataStoreIds =
             [
-                .. await connection.QueryAsync<long>(
+                .. await connection.QueryAsync<int>(
                     """
                     SELECT DataStoreId FROM dmscs.ApiClientDataStore
                     WHERE ApiClientId = @ApiClientId;
@@ -777,7 +777,7 @@ public class ApplicationRepository(
     }
 
     public async Task<ApiClientUuidSyncResult> SyncApplicationApiClientUuid(
-        long applicationId,
+        int applicationId,
         string clientId,
         Guid expectedClientUuid,
         Guid newClientUuid
