+++ src/config/backend/EdFi.DmsConfigurationService.Backend.Postgresql/Repositories/ApiClientRepository.cs
@@ -505,7 +505,7 @@ public class ApiClientRepository(
         }
     }
 
-    public async Task<ApiClientResolutionResult> GetApiClientResolutionState(long id)
+    public async Task<ApiClientResolutionResult> GetApiClientResolutionState(int id)
     {
         await using var connection = new NpgsqlConnection(databaseOptions.Value.DatabaseConnection);
         await connection.OpenAsync();
@@ -521,7 +521,7 @@ public class ApiClientRepository(
                 FOR UPDATE OF ac;
                 """;
             var client = await connection.QuerySingleOrDefaultAsync<(
-                long ApplicationId,
+                int ApplicationId,
                 string Name,
                 bool IsApproved,
                 string ClientId,
@@ -533,9 +533,9 @@ public class ApiClientRepository(
                 return new ApiClientResolutionResult.FailureNotExists();
             }
 
-            long[] dataStoreIds =
+            int[] dataStoreIds =
             [
-                .. await connection.QueryAsync<long>(
+                .. await connection.QueryAsync<int>(
                     """
                     SELECT "DataStoreId" FROM "dmscs"."ApiClientDataStore"
                     WHERE "ApiClientId" = @Id;
@@ -567,7 +567,7 @@ public class ApiClientRepository(
     }
 
     public async Task<ApiClientUuidSyncResult> SyncApiClientUuid(
-        long id,
+        int id,
         Guid expectedClientUuid,
         Guid newClientUuid
     )
